Move GitUtil tests
