+++ src/test/java/reposense/git/GitLogTest.java
@@ -1,9 +1,5 @@
 package reposense.git;
 
-import static reposense.git.GitUtil.convertToGitExcludeGlobArgs;
-import static reposense.util.StringsUtil.addQuote;
-
-import java.io.File;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.Date;
@@ -80,37 +76,6 @@ public class GitLogTest extends GitTestTemplate {
 
         String content = GitLog.getWithFiles(config, ignoreAllJavaFilesAuthor);
         Assert.assertTrue(TestUtil.compareNumberFilesChanged(0, content));
-    }
-
-    @Test
-    public void gitUtil_convertToGitExcludeGlobArgs_success() {
-        File repoRoot = new File(config.getRepoRoot());
-        final String cmdFormat = " " + addQuote(":(exclude)%s");
-        final String emptyResult = "";
-
-        String result = convertToGitExcludeGlobArgs(repoRoot, Collections.EMPTY_LIST);
-        Assert.assertEquals(result, emptyResult);
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("**.js"));
-        Assert.assertEquals(result, String.format(cmdFormat, "**.js"));
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("collate**"));
-        Assert.assertEquals(result, String.format(cmdFormat, "collate**"));
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("*\\sth"));
-        Assert.assertEquals(result, String.format(cmdFormat, "*\\sth"));
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("bin/*"));
-        Assert.assertEquals(result, String.format(cmdFormat, "bin/*"));
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("../**"));
-        Assert.assertEquals(result, emptyResult);
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("\\**"));
-        Assert.assertEquals(result, emptyResult);
-
-        result = convertToGitExcludeGlobArgs(repoRoot, Collections.singletonList("/sth/*"));
-        Assert.assertEquals(result, emptyResult);
     }
 
     @Test
